Reorganize piece representation removing unicode pieces
Unicode piece representation was likely never going to be used.  And it
was simple enough that I wouldn't mind adding it back later if needed.
But for now I really didn't like the pointless trait and unicode on
Piece and ColoredPiece.  I kept to_ascii for compatibility and added
display trait for piece types.