Super ugly cargo.toml dump which includes a high level todo list
The todo list in this file is extremely optimistic and its likely most
of the feature won't be implemented for a long time, if ever.
That being said I do plan to keep the feature list this modular to
easily modify the engine to meet any set of demands.
For now there are two collections of features: default and beta.
Default is everything that makes the engine production ready and the
most competitive.  Beta is the same as  default but includes some
developer/tester features like diagnostics and tuning.